requirements.txt
kubernetes>=29.0.0,<30.0.0
google-cloud-dns>=0.34.0,<1.0.0
google-auth>=2.0.0,<3.0.0
google-api-core>=2.0.0,<3.0.0
